divide command to more files
